refactor: refactor theme-related types
