+++ packages/console/src/components/AppError/index.tsx
@@ -1,3 +1,4 @@
+import { Theme } from '@logto/schemas';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -5,7 +6,6 @@ import ErrorDark from '@/assets/images/error-dark.svg';
 import Error from '@/assets/images/error.svg';
 import KeyboardArrowDown from '@/assets/images/keyboard-arrow-down.svg';
 import KeyboardArrowUp from '@/assets/images/keyboard-arrow-up.svg';
-import { Theme } from '@/types/theme';
 import { onKeyDownHandler } from '@/utils/a11y';
 import { getThemeFromLocalStorage } from '@/utils/theme';
