+++ packages/console/src/components/AppLoading/Offline.tsx
@@ -1,7 +1,8 @@
+import { Theme } from '@logto/schemas';
+
 import IllustrationDark from '@/assets/images/loading-illustration-dark.svg';
 import Illustration from '@/assets/images/loading-illustration.svg';
 import { Daisy as Spinner } from '@/components/Spinner';
-import { Theme } from '@/types/theme';
 import { getThemeFromLocalStorage } from '@/utils/theme';
 
 import * as styles from './index.module.scss';
