+++ packages/console/src/components/AppLoading/index.tsx
@@ -1,8 +1,9 @@
+import { Theme } from '@logto/schemas';
+
 import IllustrationDark from '@/assets/images/loading-illustration-dark.svg';
 import Illustration from '@/assets/images/loading-illustration.svg';
 import { Daisy as Spinner } from '@/components/Spinner';
 import useTheme from '@/hooks/use-theme';
-import { Theme } from '@/types/theme';
 
 import * as styles from './index.module.scss';
