+++ packages/console/src/components/ConnectorLogo/index.tsx
@@ -1,8 +1,8 @@
 import type { ConnectorResponse } from '@logto/schemas';
+import { Theme } from '@logto/schemas';
 import classNames from 'classnames';
 
 import useTheme from '@/hooks/use-theme';
-import { Theme } from '@/types/theme';
 
 import ImageWithErrorFallback from '../ImageWithErrorFallback';
 import * as styles from './index.module.scss';
