+++ packages/console/src/components/EmptyDataPlaceholder/index.tsx
@@ -1,10 +1,10 @@
+import { Theme } from '@logto/schemas';
 import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
 
 import EmptyDark from '@/assets/images/table-empty-dark.svg';
 import Empty from '@/assets/images/table-empty.svg';
 import useTheme from '@/hooks/use-theme';
-import { Theme } from '@/types/theme';
 
 import * as styles from './index.module.scss';
