+++ packages/console/src/components/ImageWithErrorFallback/index.tsx
@@ -1,10 +1,10 @@
+import { Theme } from '@logto/schemas';
 import type { ImgHTMLAttributes } from 'react';
 import { useState } from 'react';
 
 import FallbackImageDark from '@/assets/images/broken-image-dark.svg';
 import FallbackImageLight from '@/assets/images/broken-image-light.svg';
 import useTheme from '@/hooks/use-theme';
-import { Theme } from '@/types/theme';
 
 type Props = { containerClassName?: string } & ImgHTMLAttributes<HTMLImageElement>;
