+++ packages/console/src/components/SignInExperiencePreview/index.tsx
@@ -1,6 +1,6 @@
 import type { LanguageTag } from '@logto/language-kit';
 import type { ConnectorMetadata, ConnectorResponse, SignInExperience } from '@logto/schemas';
-import { ConnectorType } from '@logto/schemas';
+import { Theme, ConnectorType } from '@logto/schemas';
 import { conditional } from '@silverhand/essentials';
 import classNames from 'classnames';
 import { format } from 'date-fns';
@@ -14,13 +14,13 @@ import type { RequestError } from '@/hooks/use-api';
 import useUiLanguages from '@/hooks/use-ui-languages';
 
 import * as styles from './index.module.scss';
-import { PreviewPlatform, UiTheme } from './types';
+import { PreviewPlatform } from './types';
 
 export { default as ToggleUiThemeButton } from './components/ToggleUiThemeButton';
 
 type Props = {
   platform: PreviewPlatform;
-  mode: UiTheme;
+  mode: Theme;
   language?: LanguageTag;
   signInExperience?: SignInExperience;
 };
@@ -104,7 +104,7 @@ const SignInExperiencePreview = ({ platform, mode, language = 'en', signInExperi
       style={conditional(
         platform === PreviewPlatform.DesktopWeb && {
           // Set background color to match iframe's background color on both dark and light mode.
-          backgroundColor: mode === UiTheme.Dark ? '#000' : '#e5e1ec',
+          backgroundColor: mode === Theme.Dark ? '#000' : '#e5e1ec',
         }
       )}
     >
