+++ packages/console/src/components/SignInExperiencePreview/types.ts
@@ -1,14 +1,3 @@
-/**
- * Note:
- * The UI theme and the AdminConsole theme are unrelated.
- * The UI theme should be used to set the SignInExperience preview.
- * The definition of the UI theme can be found in `packages/ui/src/types/index.ts`.
- */
-export enum UiTheme {
-  Light = 'light',
-  Dark = 'dark',
-}
-
 export enum PreviewPlatform {
   DesktopWeb = 'desktopWeb',
   Mobile = 'mobile',
