+++ packages/console/src/components/Table/TableError.tsx
@@ -1,9 +1,9 @@
+import { Theme } from '@logto/schemas';
 import { useTranslation } from 'react-i18next';
 
 import RequestErrorDarkImage from '@/assets/images/request-error-dark.svg';
 import RequestErrorImage from '@/assets/images/request-error.svg';
 import useTheme from '@/hooks/use-theme';
-import { Theme } from '@/types/theme';
 
 import Button from '../Button';
 import * as styles from './TableError.module.scss';
