+++ packages/console/src/containers/AppContent/components/UserInfo/index.tsx
@@ -1,5 +1,6 @@
 import { builtInLanguageOptions as consoleBuiltInLanguageOptions } from '@logto/phrases';
 import { useLogto } from '@logto/react';
+import { Theme } from '@logto/schemas';
 import classNames from 'classnames';
 import { useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -18,7 +19,7 @@ import UserInfoCard from '@/components/UserInfoCard';
 import { getSignOutRedirectPathname } from '@/consts';
 import useCurrentUser from '@/hooks/use-current-user';
 import useUserPreferences from '@/hooks/use-user-preferences';
-import { Theme } from '@/types/theme';
+import { ThemePolicy } from '@/types/appearance-mode';
 import { onKeyDownHandler } from '@/utils/a11y';
 
 import SubMenu from '../SubMenu';
@@ -98,7 +99,7 @@ const UserInfo = () => {
           title="menu.appearance.label"
           options={[
             {
-              value: 'system',
+              value: ThemePolicy.System,
               title: t('menu.appearance.system'),
             },
             {
