+++ packages/console/src/onboarding/pages/SignInExperience/components/Preview/index.tsx
@@ -1,11 +1,12 @@
 import type { SignInExperience } from '@logto/schemas';
+import { Theme } from '@logto/schemas';
 import { conditional } from '@silverhand/essentials';
 import classNames from 'classnames';
 import { useState } from 'react';
 
 import LivePreviewButton from '@/components/LivePreviewButton';
 import SignInExperiencePreview from '@/components/SignInExperiencePreview';
-import { PreviewPlatform, UiTheme } from '@/components/SignInExperiencePreview/types';
+import { PreviewPlatform } from '@/components/SignInExperiencePreview/types';
 
 import PlatformTabs from '../PlatformTabs';
 import * as styles from './index.module.scss';
@@ -31,7 +32,7 @@ const Preview = ({ signInExperience, isLivePreviewDisabled = false, className }:
       </div>
       <SignInExperiencePreview
         platform={currentTab}
-        mode={UiTheme.Light}
+        mode={Theme.Light}
         signInExperience={signInExperience}
       />
     </div>
