+++ packages/console/src/pages/NotFound/index.tsx
@@ -1,10 +1,10 @@
+import { Theme } from '@logto/schemas';
 import { useTranslation } from 'react-i18next';
 
 import NotFoundDarkImage from '@/assets/images/not-found-dark.svg';
 import NotFoundImage from '@/assets/images/not-found.svg';
 import Card from '@/components/Card';
 import useTheme from '@/hooks/use-theme';
-import { Theme } from '@/types/theme';
 
 import * as styles from './index.module.scss';
