+++ packages/console/src/pages/SignInExperience/components/Preview/index.tsx
@@ -1,6 +1,7 @@
 import type { LanguageTag } from '@logto/language-kit';
 import { languages as uiLanguageNameMapping } from '@logto/language-kit';
 import type { SignInExperience } from '@logto/schemas';
+import { Theme } from '@logto/schemas';
 import classNames from 'classnames';
 import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -8,7 +9,7 @@ import { useTranslation } from 'react-i18next';
 import LivePreviewButton from '@/components/LivePreviewButton';
 import Select from '@/components/Select';
 import SignInExperiencePreview, { ToggleUiThemeButton } from '@/components/SignInExperiencePreview';
-import { PreviewPlatform, UiTheme } from '@/components/SignInExperiencePreview/types';
+import { PreviewPlatform } from '@/components/SignInExperiencePreview/types';
 import TabNav, { TabNavItem } from '@/components/TabNav';
 import useUiLanguages from '@/hooks/use-ui-languages';
 
@@ -29,13 +30,13 @@ const Preview = ({
 }: Props) => {
   const { t } = useTranslation(undefined, { keyPrefix: 'admin_console' });
   const [language, setLanguage] = useState<LanguageTag>('en');
-  const [mode, setMode] = useState<UiTheme>(UiTheme.Light);
+  const [mode, setMode] = useState<Theme>(Theme.Light);
   const [platform, setPlatform] = useState<PreviewPlatform>(PreviewPlatform.DesktopWeb);
   const { languages } = useUiLanguages();
 
   useEffect(() => {
     if (!signInExperience?.color.isDarkModeEnabled) {
-      setMode(UiTheme.Light);
+      setMode(Theme.Light);
     }
   }, [mode, signInExperience]);
