+++ packages/console/src/pages/SignInExperience/components/Welcome/index.tsx
@@ -1,3 +1,4 @@
+import { Theme } from '@logto/schemas';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -5,7 +6,6 @@ import WelcomeImageDark from '@/assets/images/sign-in-experience-welcome-dark.sv
 import WelcomeImage from '@/assets/images/sign-in-experience-welcome.svg';
 import Button from '@/components/Button';
 import useTheme from '@/hooks/use-theme';
-import { Theme } from '@/types/theme';
 
 import GuideModal from './GuideModal';
 import * as styles from './index.module.scss';
