+++ packages/schemas/src/types/index.ts
@@ -15,3 +15,4 @@ export * from './tenant.js';
 export * from './user-assets.js';
 export * from './hook.js';
 export * from './service-log.js';
+export * from './theme.js';
