+++ packages/schemas/src/types/theme.ts
@@ -1,0 +1,4 @@
+export enum Theme {
+  Light = 'light',
+  Dark = 'dark',
+}
